src/requirements.txt
astroid==2.3.3
click==7.1.1
docutils==0.16
Flask==1.1.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
MarkupSafe==1.1.1
mccabe==0.6.1
psutil==5.7.0
pyftpdlib==1.5.6
pylint==2.4.4
python-daemon==2.2.4
six==1.14.0
typed-ast==1.4.1
urwid==2.1.0
Werkzeug==1.0.1
wrapt==1.11.2
